Moved location to volunteer registration and updated UI so all fields fit on screen
