+++ app/src/main/res/layout/activity_set_skills.xml
@@ -45,32 +45,6 @@
         android:layout_below="@id/etSkill2"
         android:hint="@string/enter_skill" />
 
-    <TextView
-        android:id="@+id/tvLocation"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="5dp"
-        android:layout_marginLeft="15dp"
-        android:layout_marginRight="15dp"
-        android:layout_marginTop="5dp"
-        android:layout_below="@id/etSkill3"
-        android:text="@string/location_message"
-        android:layout_centerHorizontal="true"/>
-
-    <EditText
-        android:id="@+id/etLocation"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="5dp"
-        android:layout_marginLeft="15dp"
-        android:layout_marginRight="15dp"
-        android:layout_marginTop="5dp"
-        android:layout_below="@id/tvLocation"
-        android:hint="@string/enter_city"
-        android:focusable="false"
-        android:clickable="true"
-        android:inputType="none"/>
-
     <Button
         android:id="@+id/setSkills"
         android:layout_width="match_parent"
