+++ app/src/main/res/layout/activity_vol_reg.xml
@@ -26,7 +26,8 @@
         android:layout_marginLeft="15dp"
         android:layout_marginRight="15dp"
         android:layout_marginTop="5dp"
-        android:hint="@string/name" />
+        android:hint="@string/name"
+        android:textSize="15dp"/>
 
     <EditText
         android:id="@+id/etEmail"
@@ -39,7 +40,8 @@
         android:layout_marginRight="15dp"
         android:layout_marginTop="5dp"
         android:hint="@string/email"
-        android:inputType="textEmailAddress" />
+        android:inputType="textEmailAddress"
+        android:textSize="15dp"/>
 
     <EditText
         android:id="@+id/etPassword"
@@ -51,7 +53,8 @@
         android:layout_marginRight="15dp"
         android:layout_marginTop="5dp"
         android:hint="@string/password"
-        android:inputType="textPassword" />
+        android:inputType="textPassword"
+        android:textSize="15dp"/>
 
     <EditText
         android:id="@+id/etConfirmPassword"
@@ -63,7 +66,8 @@
         android:layout_marginBottom="5dp"
         android:layout_marginLeft="15dp"
         android:layout_marginRight="15dp"
-        android:layout_marginTop="5dp"/>
+        android:layout_marginTop="5dp"
+        android:textSize="15dp"/>
 
     <EditText
         android:id="@+id/etAge"
@@ -75,7 +79,8 @@
         android:layout_marginBottom="5dp"
         android:layout_marginLeft="15dp"
         android:layout_marginRight="15dp"
-        android:layout_marginTop="5dp"/>
+        android:layout_marginTop="5dp"
+        android:textSize="15dp"/>
 
     <EditText
         android:id="@+id/etPhone"
@@ -87,13 +92,29 @@
         android:layout_marginBottom="5dp"
         android:layout_marginLeft="15dp"
         android:layout_marginRight="15dp"
-        android:layout_marginTop="5dp"/>
+        android:layout_marginTop="5dp"
+        android:textSize="15dp"/>
+
+    <EditText
+        android:id="@+id/etLocation"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginBottom="5dp"
+        android:layout_marginLeft="15dp"
+        android:layout_marginRight="15dp"
+        android:layout_marginTop="5dp"
+        android:layout_below="@id/etPhone"
+        android:hint="@string/enter_city"
+        android:focusable="false"
+        android:clickable="true"
+        android:inputType="none"
+        android:textSize="15dp"/>
 
     <Button
         android:id="@+id/btnRegister"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_below="@id/etPhone"
+        android:layout_below="@id/etLocation"
         android:text="@string/register"
         android:layout_marginBottom="5dp"
         android:layout_marginLeft="15dp"
